+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## unreleased
 
 - feat: Crash event and session in same envelope #716
+- feat: Migrate session init for stored envelopes #693
+- fix: Remove redundant sdk options enable check in SentryHttpTransport #698
 
 ## 6.0.0-alpha.0
 
@@ -33,7 +35,6 @@ Other new features and fixes:
 - feat: Implement NSCopying for SentrySession #683
 - fix: Crash when SentryClient is nil in SentryHub #681
 - feat: Send cached envelopes first #676
-- fix: Remove redundant sdk options enable check in SentryHttpTransport #698
 
 ## 5.2.2